// src/generated/prisma/models/PredictionOption.ts
export type PredictionOptionModel = runtime.Types.Result.DefaultSelection<Prisma.$PredictionOptionPayload>
// ...
export type $PredictionOptionPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "PredictionOption"
  objects: {
    prediction: Prisma.$PredictionPayload<ExtArgs>
    votes: Prisma.$VotePayload<ExtArgs>[]
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    predictionId: string
    label: string
  }, ExtArgs["result"]["predictionOption"]>
  composites: {}
}