// src/generated/prisma/models/Vote.ts
export type VoteModel = runtime.Types.Result.DefaultSelection<Prisma.$VotePayload>
// ...
export type $VotePayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Vote"
  objects: {
    prediction: Prisma.$PredictionPayload<ExtArgs>
    option: Prisma.$PredictionOptionPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    predictionId: string
    optionId: string
    userId: string
    createdAt: Date
  }, ExtArgs["result"]["vote"]>
  composites: {}
}